Work in feedback from Jonathan.
Meeting: 12/08/2017

- I don't have to put the exact names of my college courses, nor do I
have to put a comprehensive summary of courses. They do have to reflect
things I've actually taken, but that's it.
- Given the previous point, I should re-write the course names to be as
self-descriptive as possible of the content of the course and skills
learned.
- Recommended that I put some more skills down on the resume. Since I'm
a recent college grad and employers will know this, I should put down
skills that I know some intermediate amount of knowledge in. Pretty
much, so long as I've proficiently used it for a project before, or
feasibly could. For SQL in particular, he mentioned the use of queries
and joins as "knowing SQL".
- The skills section was pretty lacking. Separated it into programming
languages and technologies. The technologies can include different
things that I know, such as Node.js or Express or SQL.
- Separated experience into projects and experience. This helps
differentiate my employment experience from the things that I've done.
- Took out "Relevant" from "Relevant Experience". All the experience
there should be relevant. Resume is not a comprehensive employment
history. There are other services for that, such as LinkedIn.
- With some of the bullets, tried to be more specific about the actual
things that I did. I orignally opted to be more vague about the specific
accomplishments and instead tried to describe a "meta" accomplishment.
For instance, instead of saying that I built a chatbot using NLTK in a
hackathon, I focused on the fact that I worked with a team of diverse
skills on short notice to create a working product. I didn't think the
product mattered, especially since the chatbot barely functioned. I
didn't learn enough about making one to really say that I'm experienced
with NLTK.