Tons of changes, major step in pollen work
The pollen.rkt outputs nearly complete products for both pdf and odt.
It's a messy piece of shit, however. Very brittle. That all needs to get
changed.

Managed to get the pdf version to make room on one line for entries with
really long job titles.

Changed the templates around a little bit.

In some sense, this branch is complete, but it needs more work. The
pollen stuff works, but not well, and I reading it is a nightmare. I can
only imagine that the same is true for editing it later on.